# Conflicts:
#	cottontaildb-dbms/src/main/kotlin/org/vitrivr/cottontail/server/grpc/services/DDLService.kt
#	cottontaildb-dbms/src/test/kotlin/org/vitrivr/cottontail/math/knn/FloatVectorDistanceTest.kt

+++ cottontaildb-core/src/main/kotlin/org/vitrivr/cottontail/core/values/ShortValue.kt
@@ -22,7 +22,7 @@ value class ShortValue(override val value: Short): RealValue<Short> {
         val MIN_VALUE = ShortValue(Short.MIN_VALUE + 1.toShort())
 
         /** The maximum value that can be held by a [ShortValue]. */
-        val MAX_VALUE = ShortValue(Int.MAX_VALUE)
+        val MAX_VALUE = ShortValue(Short.MAX_VALUE)
 
         /** The zero [ShortValue]. */
         val ZERO = ShortValue(0.toShort())
